[tests] Added a method for retrieving the package name from code.
The tests for this method are within the JavaProjectFixtureTest class.
Inserted a static field that provides a regular expression group for
Java identifiers as stated in the Java Language Specification. This
field will not only be used in the new method but also in the other
find* methods from the JavaProjectFixture class to match Java
identifiers.

Change-Id: I8f76599a627dfaa88b8ae37ac62ffc41a5a9b226