[completion] Added completion engine for extended proposal tracking and proposal manipulation.
Note, this is an highly experimental API and quite untested. There are
known bugs such as completions with existing prefixes don't work
correctly. This extension is made available because of the GSOC
project proposal.

Please study the following API classes: 

* SessionProcessor
* ProposalProcessor
* RecommendersAllCompletionEngine

and their various implementations in the sandbox project to get a
feeling what these classes are about.


Change-Id: Ib5b7efdf19e127cc2cea98b0e28e3d9b837bf520