[models] Add JREDirectoryNameAdvisor
Add JREDirectoryNameAdvisor which try to detect the version of the jre out
of the folder name. This advisor is needed to detect e.g jdk 1.6 which
is not added as Execution Environment (no other jre advisor fits to this
case).

The advisor may get confused by non-canonical path names and also by the
fact that an 1.7.0 JRE is sometimes called Java 7.

Change-Id: Ieb14831275b715ef11dd09c9b1af1c540707242c
Signed-off-by: Olav Lenz <olav.lenz@gmail.com>