[snipmatch] Bug 436587: Introduce ISnippetRepositoryConfiguration.
Changed the way snippet repositories are created. All needed infos are
stored in configurations (ISnippetRepositoryConfiguration) and stored in
the preferences. Providers (ISnippetRepositoryProviders) are able to
create concrete ISnippetRepository instance for a given configuration.
The snippet view and preference page are able to work with the setup.
This change might break thrid-party plugins because of the new extension
point mechanism.

Change-Id: I840e0ae32d61ad26bb1a951f682ae77e1be336ca
Signed-off-by: Olav Lenz <olav.lenz@gmail.com>