[snipmatch] Bug 438897: Separate and adjustable Fetch & Push Urls
This changes the fetch Url from /r/recommenders/org...snippets to
/gitroot/recommenders/org...snippets.git
The push Url can be changed by the user and is the /r/recommenders/org...snippets one by default.
The branch prefix can be changed as well and is "refs/for" by default, which is appended with our format string, i.e. "format-3"

Change-Id: Ie326fe59ae7e494da19dd25446a9cb51441d0722
Signed-off-by: Johannes Dorn <johannes.dorn@codetrails.com>