[snipmatch] Bug 438164: Repo need not be open *before* "Create Snippet"
It has, however, to be open by the time the user saves the new snippet.
At the moment, the error reporting is not great: The user is presented
with a "Save Failed: IllegalStateException" dialog. The editor, however,
remains open and the user can save again; this will succeed once the
repo has been opened in the background.

Change-Id: Iffca8bbaf5d7cbf1ada1fdd003f336a12a731ff3
Signed-off-by: Andreas Sewe <andreas.sewe@codetrails.com>