[completion] Bug 442625: tips processor consumes a lot of time
the tips processor accesses the list of available proposals to determine
whether it would prevent autocompletion. This access triggers a the
available proposas to be computed - which in turn consumes some time. 

This change is not really a fix. It merely puts the check whether it has
proposals to offer first.

Change-Id: Iac7dc69377546187b10c5ab792ed6e7a149dfc05